requirements.txt
# Visual FaQtory v0.9.0-beta — Requirements
# ═══════════════════════════════════════════════════════════════════════════════

# Core (always needed)
pyyaml>=6.0

# Backend: ComfyUI API
requests>=2.28
websocket-client>=1.4

# Backend: Google Veo (optional — only needed for backend.type=veo)
google-genai>=1.0

# Image processing (used by mock backend, image metrics)
pillow>=9.0
numpy>=1.24

# Crowd Control (web prompt page + QR overlay for live streams)
fastapi>=0.110
uvicorn[standard]>=0.27
qrcode>=7.4
obsws_python

# other
scipy>=1.17.1
